Переименовать AuthorizedCourse в Workout;
