+++ src/pages/Workout/Progress/ProgressBar.module.scss
