+++ src/routes/routes.jsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom';
 import Main from '../pages/Main/Main';
 import Profile from '../pages/Profile/Profile';
 import Course from '../pages/Course/Course';
-import { AuthorizedCourse } from '../pages/authorizedCourse/AuthorizedCourse';
+import { Workout } from '../pages/Workout/Workout';
 import ProtectedRoute from './ProtectedRoute';
 import { Auth } from '../pages/Auth/Auth';
 
@@ -19,7 +19,7 @@ const AppRoutes = () => {
 
 			<Route element={<ProtectedRoute />}>
 				<Route path="/profile" element={<Profile />} />
-				<Route path="/workout/:id" element={<AuthorizedCourse />} />
+				<Route path="/workout/:id" element={<Workout />} />
 			</Route>
 		</Routes>
 	);
